Move IndifferentJson into Store namespace, add config flag
Address review feedback from @byroot:

1. Rename Type::IndifferentJson to Store::IndifferentJsonType to make
   it clear the type is exclusive to AR::Store (it only handles Hash
   top-level objects, which store enforces).

2. Add config.active_record.store_native_json_columns flag (defaults
   to false, set to true in load_defaults "8.2") so existing apps
   with double-serialized data in json columns can keep the old
   behavior until they migrate their data.
