+++ activerecord/lib/active_record/type.rb
@@ -8,7 +8,6 @@ require "active_record/type/date"
 require "active_record/type/date_time"
 require "active_record/type/decimal_without_scale"
 require "active_record/type/json"
-require "active_record/type/indifferent_json"
 require "active_record/type/time"
 require "active_record/type/text"
 require "active_record/type/unsigned_integer"
