+++ railties/lib/rails/application/configuration.rb
@@ -388,6 +388,7 @@ module Rails
           if respond_to?(:active_record)
             active_record.postgresql_adapter_decode_bytea = true
             active_record.postgresql_adapter_decode_money = true
+            active_record.store_native_json_columns = true
           end
 
           if respond_to?(:active_storage)
